Move AppCore and AppPlugins interfaces back to shim and re-export them from app types
